move cors to tools
