+++ lib/tools.js
@@ -14,6 +14,7 @@ let _ = require('./translate')._;
 let util = require('util');
 let createDOMPurify = require('dompurify');
 let htmlToText = require('html-to-text');
+let config = require('config');
 
 let blockedUsers = ['abuse', 'admin', 'billing', 'compliance', 'devnull', 'dns', 'ftp', 'hostmaster', 'inoc', 'ispfeedback', 'ispsupport', 'listrequest', 'list', 'maildaemon', 'noc', 'noreply', 'noreply', 'null', 'phish', 'phishing', 'postmaster', 'privacy', 'registrar', 'root', 'security', 'spam', 'support', 'sysadmin', 'tech', 'undisclosedrecipients', 'unsubscribe', 'usenet', 'uucp', 'webmaster', 'www'];
 
@@ -30,6 +31,7 @@ module.exports = {
     prepareHtml,
     purifyHTML,
     mergeTemplateIntoLayout,
+    getCorsOptions,
     workers: new Set()
 };
 
@@ -318,3 +320,24 @@ function mergeTemplateIntoLayout(template, layout, callback) {
         return done(template, layout);
     }
 }
+
+    function getCorsOptions() {
+        let originWhitelist = config.cors && config.cors.origins || [];
+
+        let corsOptions = {
+            allowedHeaders: ['Content-Type', 'Origin', 'Accept', 'X-Requested-With'],
+            methods: ['GET', 'POST'],
+            optionsSuccessStatus: 200, // IE11 chokes on 204
+            origin: (origin, callback) => {
+                if (originWhitelist.includes(origin)) {
+                    callback(null, true);
+                } else {
+                    let err = new Error(_('Not allowed by CORS'));
+                    err.status = 403;
+                    callback(err);
+                }
+            }
+        };
+
+        return corsOptions;
+    }
