+++ routes/api.js
@@ -9,8 +9,11 @@ let confirmations = require('../lib/models/confirmations');
 let tools = require('../lib/tools');
 let express = require('express');
 let log = require('npmlog');
+let cors = require('cors');
 let router = new express.Router();
 let mailHelpers = require('../lib/subscription-mail-helpers');
+
+router.use(cors(tools.getCorsOptions));
 
 router.all('/*', (req, res, next) => {
     if (!req.query.access_token) {
