+++ routes/subscription.js
@@ -20,26 +20,9 @@ let geoip = require('geoip-ultralight');
 let confirmations = require('../lib/models/confirmations');
 let mailHelpers = require('../lib/subscription-mail-helpers');
 
-let originWhitelist = config.cors && config.cors.origins || [];
-
-let corsOptions = {
-    allowedHeaders: ['Content-Type', 'Origin', 'Accept', 'X-Requested-With'],
-    methods: ['GET', 'POST'],
-    optionsSuccessStatus: 200, // IE11 chokes on 204
-    origin: (origin, callback) => {
-        if (originWhitelist.includes(origin)) {
-            callback(null, true);
-        } else {
-            let err = new Error(_('Not allowed by CORS'));
-            err.status = 403;
-            callback(err);
-        }
-    }
-};
-
 let corsOrCsrfProtection = (req, res, next) => {
     if (req.get('X-Requested-With') === 'XMLHttpRequest') {
-        cors(corsOptions)(req, res, next);
+        cors(tools.getCorsOptions)(req, res, next);
     } else {
         passport.csrfProtection(req, res, next);
     }
